Default to Float32 eltype for prob

+++ src/preferences.jl
@@ -5,7 +5,7 @@ function default_ci_fun()
     return ci_kind == :eti ? eti : hdi
 end
 
-function default_ci_prob((::Type{T})=Float64) where {T<:Real}
+function default_ci_prob(::Type{T}=Float32) where {T<:Real}
     prob = T(Preferences.load_preference(PosteriorStats, "ci_prob", 0.94))
     0 < prob < 1 || throw(DomainError(prob, "ci_prob must be in the range (0, 1)."))
     return prob
